+++ .gitignore
@@ -1,3 +1,5 @@
 .DS_Store
 
 config
+node_modules
+dist
