+++ scripts/FormValidator.js
@@ -1,6 +1,6 @@
 "use strict";
 
-class FormValidator {
+export class FormValidator {
   _errorMessages = {
     empty: "Это обязательное поле",
     wrongLength: "Должно быть от 2 до 30 символов",
