+++ scripts/UserInfo.js
@@ -1,6 +1,6 @@
 "use strict";
 
-class UserInfo {
+export class UserInfo {
   constructor(name, about) {
     this._name = name;
     this._about = about;
